Modification of code structure
