+++ main.py
@@ -1,5 +1,6 @@
-import scripts.snp_charts as snp
+from scripts import snp_charts as snp
+import sys
 
 if __name__ == '__main__':
-    import sys
     snp.main_using_getopts(sys.argv[1:])
+
