+++ scripts/__init__.py
@@ -1,0 +1,2 @@
+from . import getopts_parser
+from . import utilities
