+++ scripts/utilities/__init__.py
@@ -1,0 +1,2 @@
+from . import tests
+from . import utilities
